Move classes used during execution into dedicated package
