+++ src/main/php/test/Args.class.php
@@ -1,6 +1,7 @@
 <?php namespace test;
 
 use lang\IllegalArgumentException;
+use test\execution\Context;
 
 /**
  * Selects command line arguments
@@ -30,6 +31,7 @@ class Args implements Provider {
    * @return iterable
    */
   public function values($context) {
+    var_dump($context);
 
     // Select all arguments
     if (empty($this->select)) {
