+++ src/main/php/test/execution/GroupFailed.class.php
@@ -1,4 +1,4 @@
-<?php namespace test;
+<?php namespace test\execution;
 
 use lang\{Throwable, XPException};
